+++ R/surv-brier_survival.R
@@ -22,7 +22,11 @@
 #' predicting with {censored} model. This should be an unquoted column name
 #' although this argument is passed by expression and supports
 #' [quasiquotation][rlang::quasiquotation] (you can unquote column names). For
-#' `_vec()` functions, a numeric vector. See Details below.
+#' `_vec()` functions, the dots are not used.
+#' 
+#' @param estimate A list column of data.frames corresponding to the output 
+#' given when predicting with {censored} model. See the details for more 
+#' information regarding format.
 #'
 #' @details
 #'